requirements.txt
scrapfly-sdk
python-dotenv
parsel
